requirements.txt
django
django-crispy-forms

requests

pylint
pylint-django
